+++ android/capacitor/src/main/java/com/getcapacitor/plugin/util/HttpRequestHandler.java
@@ -302,14 +302,18 @@ public class HttpRequestHandler {
             if ("null".equals(input.trim())) {
                 return JSONObject.NULL;
             } else if ("true".equals(input.trim())) {
-                return new JSONObject().put("flag", "true");
+                return true;
             } else if ("false".equals(input.trim())) {
-                return new JSONObject().put("flag", "false");
+                return false;
             } else if (input.trim().length() <= 0) {
                 return "";
             } else if (input.trim().matches("^\".*\"$")) {
                 // a string enclosed in " " is a json value, return the string without the quotes
                 return input.trim().substring(1, input.trim().length() - 1);
+            } else if (input.trim().matches("^-?\\d+$")) {
+                return Integer.parseInt(input.trim());
+            } else if (input.trim().matches("^-?\\d+(\\.\\d+)?$")) {
+                return Double.parseDouble(input.trim());
             } else {
                 try {
                     return new JSObject(input);
@@ -318,7 +322,7 @@ public class HttpRequestHandler {
                 }
             }
         } catch (JSONException e) {
-            return new JSArray(input);
+            return input;
         }
     }
 
@@ -411,16 +415,22 @@ public class HttpRequestHandler {
             }
         }
 
+        call.getData().put("activeCapacitorHttpUrlConnection", connection);
         connection.connect();
 
-        return buildResponse(connection, responseType);
+        JSObject response = buildResponse(connection, responseType);
+
+        connection.disconnect();
+        call.getData().remove("activeCapacitorHttpUrlConnection");
+
+        return response;
     }
 
     private static Boolean isDomainExcludedFromSSL(Bridge bridge, URL url) {
         try {
             Class<?> sslPinningImpl = Class.forName("io.ionic.sslpinning.SSLPinning");
             Method method = sslPinningImpl.getDeclaredMethod("isDomainExcluded", Bridge.class, URL.class);
-            return (Boolean) method.invoke(sslPinningImpl.newInstance(), bridge, url);
+            return (Boolean) method.invoke(sslPinningImpl.getDeclaredConstructor().newInstance(), bridge, url);
         } catch (Exception ignored) {
             return false;
         }