+++ cli/src/config.ts
@@ -24,6 +24,7 @@ import { fatal, isFatal } from './errors';
 import { logger } from './log';
 import { tryFn } from './util/fn';
 import { formatJSObject } from './util/js';
+import { findNXMonorepoRoot, isNXMonorepo } from './util/monorepotools';
 import { requireTS, resolveNode } from './util/node';
 import { lazy } from './util/promise';
 import { getCommandOutput } from './util/subprocess';
@@ -38,6 +39,25 @@ export async function loadConfig(): Promise<Config> {
   const appRootDir = process.cwd();
   const cliRootDir = dirname(__dirname);
   const conf = await loadExtConfig(appRootDir);
+
+  const depsForNx = await (async (): Promise<
+    { devDependencies: any; dependencies: any } | object
+  > => {
+    if (isNXMonorepo(appRootDir)) {
+      const rootOfNXMonorepo = findNXMonorepoRoot(appRootDir);
+      const pkgJSONOfMonorepoRoot: any = await tryFn(
+        readJSON,
+        resolve(rootOfNXMonorepo, 'package.json'),
+      );
+      const devDependencies = pkgJSONOfMonorepoRoot?.devDependencies ?? {};
+      const dependencies = pkgJSONOfMonorepoRoot?.dependencies ?? {};
+      return {
+        devDependencies,
+        dependencies,
+      };
+    }
+    return {};
+  })();
 
   const appId = conf.extConfig.appId ?? '';
   const appName = conf.extConfig.appName ?? '';
@@ -58,6 +78,7 @@ export async function loadConfig(): Promise<Config> {
       package: (await tryFn(readJSON, resolve(appRootDir, 'package.json'))) ?? {
         name: appName,
         version: '1.0.0',
+        ...depsForNx,
       },
       ...conf,
     },
@@ -110,7 +131,9 @@ async function loadExtConfigTS(
 
     const ts = require(tsPath); // eslint-disable-line @typescript-eslint/no-var-requires
     const extConfigObject = requireTS(ts, extConfigFilePath) as any;
-    const extConfig = extConfigObject.default ?? extConfigObject;
+    const extConfig = extConfigObject.default
+      ? await extConfigObject.default
+      : extConfigObject;
 
     return {
       extConfigType: 'ts',
@@ -137,7 +160,7 @@ async function loadExtConfigJS(
       extConfigType: 'js',
       extConfigName,
       extConfigFilePath: extConfigFilePath,
-      extConfig: require(extConfigFilePath),
+      extConfig: await require(extConfigFilePath),
     };
   } catch (e: any) {
     fatal(`Parsing ${c.strong(extConfigName)} failed.\n\n${e.stack ?? e}`);
@@ -170,7 +193,7 @@ async function loadExtConfig(rootDir: string): Promise<ExtConfigPairs> {
 async function loadCLIConfig(rootDir: string): Promise<CLIConfig> {
   const assetsDir = 'assets';
   const assetsDirAbs = join(rootDir, assetsDir);
-  const iosPlatformTemplateArchive = 'ios-template.tar.gz';
+  const iosPlatformTemplateArchive = 'ios-pods-template.tar.gz';
   const iosCordovaPluginsTemplateArchive =
     'capacitor-cordova-ios-plugins.tar.gz';
   const androidPlatformTemplateArchive = 'android-template.tar.gz';
@@ -238,6 +261,10 @@ async function loadAndroidConfig(
   const buildOptions = {
     keystorePath: extConfig.android?.buildOptions?.keystorePath,
     keystorePassword: extConfig.android?.buildOptions?.keystorePassword,
+    keystoreAlias: extConfig.android?.buildOptions?.keystoreAlias,
+    keystoreAliasPassword:
+      extConfig.android?.buildOptions?.keystoreAliasPassword,
+    signingType: extConfig.android?.buildOptions?.signingType,
     releaseType: extConfig.android?.buildOptions?.releaseType,
   };
 
@@ -474,7 +501,9 @@ async function determineGemfileOrCocoapodPath(
     if (!gemfileText) {
       return 'pod';
     }
-    const cocoapodsInGemfile = new RegExp(/gem 'cocoapods'/).test(gemfileText);
+    const cocoapodsInGemfile = new RegExp(/gem\s+['"]cocoapods/).test(
+      gemfileText,
+    );
 
     if (cocoapodsInGemfile) {
       return 'bundle exec pod';
@@ -488,7 +517,7 @@ async function determineGemfileOrCocoapodPath(
 
 function formatConfigTS(extConfig: ExternalConfig): string {
   // TODO: <reference> tags
-  return `import { CapacitorConfig } from '@capacitor/cli';
+  return `import type { CapacitorConfig } from '@capacitor/cli';
 
 const config: CapacitorConfig = ${formatJSObject(extConfig)};