+++ ios/Capacitor/Capacitor/CAPInstanceDescriptor.swift
@@ -1,8 +1,8 @@
 import Foundation
 
 public enum InstanceDescriptorDefaults {
-    static let scheme = "capacitor"
-    static let hostname = "localhost"
+    public static let scheme = "capacitor"
+    public static let hostname = "localhost"
 }
 
 private extension InstanceLoggingBehavior {
@@ -33,7 +33,7 @@ internal extension InstanceDescriptor {
         // sanity check that the app directory is valid
         var isDirectory: ObjCBool = ObjCBool(false)
         if warnings.contains(.missingAppDir) == false,
-           (FileManager.default.fileExists(atPath: appLocation.path, isDirectory: &isDirectory) == false || isDirectory.boolValue == false) {
+           FileManager.default.fileExists(atPath: appLocation.path, isDirectory: &isDirectory) == false || isDirectory.boolValue == false {
             warnings.update(with: .missingAppDir)
         }
 
@@ -113,6 +113,9 @@ internal extension InstanceDescriptor {
             }
             if let scrollEnabled = config[keyPath: "ios.scrollEnabled"] as? Bool {
                 scrollingEnabled = scrollEnabled
+            }
+            if let zoomEnabled = (config[keyPath: "ios.zoomEnabled"] as? Bool) ?? (config[keyPath: "zoomEnabled"] as? Bool) {
+                zoomingEnabled = zoomEnabled
             }
             if let pluginConfig = config[keyPath: "plugins"] as? JSObject {
                 pluginConfigurations = pluginConfig