+++ ios/Capacitor/Capacitor/CAPPluginCall.swift
@@ -77,3 +77,36 @@ extension CAPPluginCall: JSValueContainer {
         errorHandler(CAPPluginCallError(message: message, code: "UNAVAILABLE", error: nil, data: [:]))
     }
 }
+
+// MARK: Codable Support
+public extension CAPPluginCall {
+    /// Encodes the given value to a ``JSObject`` and resolves the call. If an error is thrown during encoding, ``reject(_:_:_:_:)`` is called.
+    /// - Parameters:
+    ///   - data: The value to encode
+    ///   - encoder: The encoder to use. Defaults to `JSValueEncoder()`
+    ///   - messageForRejectionFromError: A closure that takes the error thrown from ``JSValueEncoder/encodeJSObject(_:)``
+    ///   and returns a string to be provided to ``reject(_:_:_:_:)``. Defaults to a function that returns "Failed encoding response".
+    func resolve<T: Encodable>(
+        with data: T,
+        encoder: JSValueEncoder = JSValueEncoder(),
+        messageForRejectionFromError: (Error) -> String = { _ in "Failed encoding response" }
+    ) {
+        do {
+            let encoded = try encoder.encodeJSObject(data)
+            resolve(encoded)
+        } catch {
+            let message = messageForRejectionFromError(error)
+            reject(message, nil, error)
+        }
+    }
+
+    /// Decodes the options to the given type.
+    /// - Parameters:
+    ///   - type: The type to decode to.
+    ///   - decoder: The decoder to use. Defaults to `JSValueDecoder()`.
+    /// - Throws: If the options cannot be decoded.
+    /// - Returns: The decoded value.
+    func decode<T: Decodable>(_ type: T.Type, decoder: JSValueDecoder = JSValueDecoder()) throws -> T {
+        try decoder.decode(type, from: options as? JSObject ?? [:])
+    }
+}